Refactor: fetchNextEvent method updated
